move timebar interactions to hook to simplify component
